Refactor to use atomic bool stop flag
Instead of stopping the thread pool from the signal handler, an atomic
boolean flag is set to make the main loop exit.

Since the tcp listener incoming() method blocks until a new connection
is opened, trigger a new iteration of the loop by connecting to the
socket from the handler that will cause the loop to break because of the
stop flag that has been set to true.

Remove the explicit stop() in the thread pool as now gets runned
automatically by the deconstructor.